requirements.txt
requests>=2.28.0
argparse>=1.4.0
python-dotenv>=0.20.0
tqdm>=4.64.0
colorama>=0.4.5
zapcli>=0.10.0
python-owasp-zap-v2.4>=0.0.18
readline>=6.2.4.1; platform_system != "Windows"
pyreadline3>=3.4.1; platform_system == "Windows"
ipaddress>=1.0.23
dnspython>=2.2.1
beautifulsoup4>=4.11.1
jinja2>=3.1.2
lxml>=4.9.1
tabulate>=0.9.0
termcolor>=2.0.1
pycurl>=7.45.1; platform_system != "Windows"
ipwhois>=1.2.0
python-whois>=0.8.0
cryptography>=37.0.4
paramiko>=2.11.0
packaging>=21.3
aiodns>=3.0.0
aiohttp>=3.8.3
cchardet>=2.1.7
pathlib>=1.0.1
typing-extensions>=4.0.0 